the fixture is placed in a separate file
