+++ lab_5/test_add_functions.py
@@ -1,25 +1,8 @@
-import os
 import pytest
-import tempfile
 
+from fixture_file import temp_data_file
 from modules.add_functions import read_data, growth
-from typing import Generator, List
-
-
-@pytest.fixture
-def temp_data_file() -> Generator[str, None, None]:
-    """
-    Creates a temporary file with test data and removes it after use.
-
-    Returns:
-        str: Path to the temporarily created file.
-    """
-    data = "2024-01-21,-5,752,З,2\n2024-01-22,-13,764,С,1\n2024-01-23,-13,765,З,1\n"
-    with tempfile.NamedTemporaryFile(delete=False, mode='w', encoding='utf-8') as f:
-        f.write(data)
-        f.flush()
-        yield f.name
-    os.remove(f.name)
+from typing import List
 
 
 def test_read_data(temp_data_file: str) -> None:
